requirements.txt
alembic==0.9.5
aniso8601==1.3.0
astroid==1.5.3
awscli==1.14.6
backports.functools-lru-cache==1.4
botocore==1.8.10
click==6.7
colorama==0.3.7
configparser==3.5.0
docutils==0.14
enum34==1.1.6
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
Flask-Testing==0.6.2
futures==3.2.0
gunicorn==19.7.1
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==3.0.0b3
mccabe==0.6.1
numpy==1.13.1
pbr==3.0.1
pep8==1.7.0
psycopg2==2.7.3.1
py==1.4.34
pyasn1==0.4.2
PyJWT==1.4.2
pylint==1.7.2
pytest==3.1.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
selenium==3.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.14
stevedore==1.23.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.12.2
wrapt==1.10.10
